test svg values in testIcon
